// src/lib/models/CampagneAvecCandidatDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CampagneDTO } from './CampagneDTO';
import type { CandidatDTO } from './CandidatDTO';
export type CampagneAvecCandidatDTO = {
    campagne?: CampagneDTO;
    candidat?: CandidatDTO;
};